Fix "log" command when crash is started with "--minimal" option
The log option '-T' is used to print the log timestamp in human-readable
form. However, this option does not work when crash is invoked with the
'--minimal' mode, and if tried, crash will spin at 100% and continuously
crash at a divide by 0 because machdep->hz == 0.

Fix this by disallowing this option in minimal mode. In addition, only
calculate the logic to calculate kt->boot_date.tv_sec when this option
is enabled.

 [ lijiang: modify the warning output message and log. ]

Fixes: c86250bce29f ("Introduction of the "log -T" option...")
Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>
Tested-by: Mathias Krause <minipli@grsecurity.net>
Reviewed-by: Wang Long <w@laoqinren.net>
Reviewed-by: Lianbo Jiang <lijiang@redhat.com>